componentize alert for failed samples
